+++ crates/nfs3_server/src/nfs_handlers.rs
@@ -4,10 +4,11 @@ use std::io::{Read, Write};
 
 use nfs3_types::nfs3::*;
 use nfs3_types::rpc::*;
-use nfs3_types::xdr_codec::{BoundedList, List, Opaque, Pack, PackedSize, Unpack};
+use nfs3_types::xdr_codec::{BoundedList, Opaque, Pack, PackedSize, Unpack};
 use tracing::{debug, error, trace, warn};
 
 use crate::context::RPCContext;
+use crate::nfs_ext::{BoundedEntryPlusList, CookieVerfExt};
 use crate::rpc::*;
 use crate::vfs::VFSCapabilities;
 
@@ -1376,68 +1377,4 @@ pub async fn nfsproc3_readlink(
         }
     }
     Ok(())
-}
-
-trait CookieVerfExt {
-    const NONE_COOKIE_VERF: cookieverf3 = cookieverf3(0u64.to_be_bytes());
-    const SOME_COOKIE_VERF: cookieverf3 = cookieverf3(0xFFCC_FFCC_FFCC_FFCCu64.to_be_bytes());
-
-    fn from_attr(dir_attr: &post_op_attr) -> Self;
-    fn is_none(&self) -> bool;
-    fn is_some(&self) -> bool;
-}
-
-impl CookieVerfExt for cookieverf3 {
-    fn from_attr(dir_attr: &post_op_attr) -> Self {
-        if let post_op_attr::Some(attr) = dir_attr {
-            let cvf_version = ((attr.mtime.seconds as u64) << 32) | (attr.mtime.nseconds as u64);
-            Self(cvf_version.to_be_bytes())
-        } else {
-            Self::SOME_COOKIE_VERF
-        }
-    }
-
-    fn is_none(&self) -> bool {
-        self == &Self::NONE_COOKIE_VERF
-    }
-
-    fn is_some(&self) -> bool {
-        !self.is_none()
-    }
-}
-
-struct BoundedEntryPlusList {
-    entries: BoundedList<entryplus3<'static>>,
-    dircount: usize,
-    accumulated_dircount: usize,
-}
-
-impl BoundedEntryPlusList {
-    fn new(dircount: usize, maxcount: usize) -> Self {
-        Self {
-            entries: BoundedList::new(maxcount),
-            dircount,
-            accumulated_dircount: 0,
-        }
-    }
-
-    fn try_push(&mut self, entry: entryplus3<'static>) -> Result<(), entryplus3> {
-        let added_dircount = size_of::<fileid3>() // fileid
-            + size_of::<u32>() + entry.name.len() // name
-            + size_of::<cookie3>(); // cookie
-
-        if self.accumulated_dircount + added_dircount > self.dircount {
-            return Err(entry);
-        }
-
-        let result = self.entries.try_push(entry);
-        if result.is_ok() {
-            self.dircount += added_dircount;
-        }
-        result
-    }
-
-    fn into_inner(self) -> List<entryplus3<'static>> {
-        self.entries.into_inner()
-    }
 }
